making confirmAppointment a new component
