+++ client/src/components/scheduler/ConfirmAppointment.js
@@ -1,0 +1,131 @@
+import React from "react";
+import { Grid, makeStyles } from "@material-ui/core";
+import { MuiPickersUtilsProvider, Calendar } from "@material-ui/pickers";
+import DateFnsUtils from "@date-io/date-fns";
+import PropTypes from "prop-types";
+
+const ConfirmAppointment = ({ appointmentDetails, setAppointmentDetails }) => {
+  const classes = useStyles();
+  const handleFormChange = (event) => {
+    const { name, value } = event.target;
+    setAppointmentDetails({ ...appointmentDetails, [name]: value });
+    console.log(appointmentDetails);
+  };
+
+  const createAppointment = () => {
+    axios.post("/api/appointment", appointmentDetails).then((res) => {
+      setAppointmentDetails({
+        eventId: "",
+        name: "",
+        email: "",
+        notes: "",
+        time: false,
+        timezone: "UTC",
+      });
+    });
+  };
+  return (
+    <>
+      <Button
+        onClick={() =>
+          setAppointmentDetails({ ...appointmentDetails, time: false })
+        }
+      >
+        Back
+      </Button>
+      <Grid
+        direction="row"
+        alignItems="center"
+        container
+        justify="flex-start"
+        className={classes.inputRow}
+      >
+        <Grid xs="2" item>
+          <InputLabel className={classes.label}>Name</InputLabel>
+        </Grid>
+
+        <Grid xs="10" item>
+          <TextField
+            variant="outlined"
+            margin="normal"
+            name="name"
+            type="text"
+            onChange={handleFormChange}
+            fullWidth
+            value={appointmentDetails.name}
+          />
+        </Grid>
+      </Grid>
+      <Grid
+        direction="row"
+        alignItems="center"
+        container
+        justify="flex-start"
+        className={classes.inputRow}
+      >
+        <Grid xs="2" item>
+          <InputLabel className={classes.label}>Email</InputLabel>
+        </Grid>
+
+        <Grid xs="10" item>
+          <TextField
+            variant="outlined"
+            margin="normal"
+            name="email"
+            type="text"
+            onChange={handleFormChange}
+            fullWidth
+            value={appointmentDetails.email}
+          />
+        </Grid>
+      </Grid>
+      <Grid
+        direction="row"
+        alignItems="center"
+        container
+        justify="flex-start"
+        className={classes.inputRow}
+      >
+        <Grid xs="2" item>
+          <InputLabel className={classes.descriptionLabel}>Notes</InputLabel>
+        </Grid>
+
+        <Grid xs="10" item>
+          <TextField
+            variant="outlined"
+            margin="normal"
+            name="notes"
+            multiline
+            onChange={handleFormChange}
+            fullWidth
+            value={appointmentDetails.notes}
+            placeholder={`Write any details that will help prepare for our meeting.
+                                    
+                                    
+                                    `}
+            className={classes.textArea}
+          />
+        </Grid>
+      </Grid>
+
+      <Button
+        color="primary"
+        className={classes.button}
+        onClick={createAppointment}
+      >
+        Schedule Event
+      </Button>
+    </>
+  );
+};
+
+const useStyles = makeStyles(() => ({
+  root: { height: "100%", overflow: "hidden" },
+}));
+
+DisplayCalendar.propTypes = {
+  selectedDate: PropTypes.object,
+  setSelectedDate: PropTypes.func,
+};
+
+export default ConfirmAppointment;
