+++ client/src/pages/Scheduler.js
@@ -12,6 +12,7 @@ import {
 import Overview from "../components/scheduler/Overview";
 import PickDate from "../components/scheduler/PickDate";
 import PickTime from "../components/scheduler/PickTime";
+import ConfirmAppointment from "../components/scheduler/ConfirmAppointment";
 import axios from "axios";
 
 const Scheduler = () => {
@@ -30,25 +31,6 @@ const Scheduler = () => {
     time: false,
     timezone: "UTC",
   });
-
-  const handleFormChange = (event) => {
-    const { name, value } = event.target;
-    setAppointmentDetails({ ...appointmentDetails, [name]: value });
-    console.log(appointmentDetails);
-  };
-
-  const createAppointment = () => {
-    axios.post("/api/appointment", appointmentDetails).then((res) => {
-      setAppointmentDetails({
-        eventId: "",
-        name: "",
-        email: "",
-        notes: "",
-        time: false,
-        timezone: "UTC",
-      });
-    });
-  };
 
   useEffect(() => {
     let queryURL = `/api/event_details/${window.location.pathname.slice(6)}`;
@@ -86,99 +68,7 @@ const Scheduler = () => {
           spacing={2}
         >
           {appointmentDetails.time ? (
-            <>
-              <Button
-                onClick={() =>
-                  setAppointmentDetails({ ...appointmentDetails, time: false })
-                }
-              >
-                Back
-              </Button>
-              <Grid
-                direction="row"
-                alignItems="center"
-                container
-                justify="flex-start"
-                className={classes.inputRow}
-              >
-                <Grid xs="2" item>
-                  <InputLabel className={classes.label}>Name</InputLabel>
-                </Grid>
-
-                <Grid xs="10" item>
-                  <TextField
-                    variant="outlined"
-                    margin="normal"
-                    name="name"
-                    type="text"
-                    onChange={handleFormChange}
-                    fullWidth
-                    value={appointmentDetails.name}
-                  />
-                </Grid>
-              </Grid>
-              <Grid
-                direction="row"
-                alignItems="center"
-                container
-                justify="flex-start"
-                className={classes.inputRow}
-              >
-                <Grid xs="2" item>
-                  <InputLabel className={classes.label}>Email</InputLabel>
-                </Grid>
-
-                <Grid xs="10" item>
-                  <TextField
-                    variant="outlined"
-                    margin="normal"
-                    name="email"
-                    type="text"
-                    onChange={handleFormChange}
-                    fullWidth
-                    value={appointmentDetails.email}
-                  />
-                </Grid>
-              </Grid>
-              <Grid
-                direction="row"
-                alignItems="center"
-                container
-                justify="flex-start"
-                className={classes.inputRow}
-              >
-                <Grid xs="2" item>
-                  <InputLabel className={classes.descriptionLabel}>
-                    Notes
-                  </InputLabel>
-                </Grid>
-
-                <Grid xs="10" item>
-                  <TextField
-                    variant="outlined"
-                    margin="normal"
-                    name="notes"
-                    multiline
-                    onChange={handleFormChange}
-                    fullWidth
-                    value={appointmentDetails.notes}
-                    placeholder={`Write any details that will help prepare for our meeting.
-                                    
-                                    
-                                    `}
-                    className={classes.textArea}
-                  />
-                </Grid>
-              </Grid>
-
-              <Button
-                color="primary"
-                className={classes.button}
-                onClick={createAppointment}
-              >
-                Schedule Event
-              </Button>
-            </>
+            <ConfirmAppointment />
           ) : (
             <>
               <Grid item>
